enable wconversion and finish fixes (#3728)
* enable wconversion and finish fixes

* don't pass -Wconversion if it's win32

+++ src/_cffi_src/openssl/ecdh.py
@@ -19,7 +19,7 @@ FUNCTIONS = """
 MACROS = """
 int ECDH_compute_key(void *, size_t, const EC_POINT *, EC_KEY *,
                      void *(*)(const void *, size_t, void *, size_t *));
-int SSL_CTX_set_ecdh_auto(SSL_CTX *, int);
+long SSL_CTX_set_ecdh_auto(SSL_CTX *, int);
 """
 
 CUSTOMIZATIONS = """
@@ -27,7 +27,7 @@ static const long Cryptography_HAS_ECDH = 1;
 
 #ifndef SSL_CTX_set_ecdh_auto
 static const long Cryptography_HAS_SET_ECDH_AUTO = 0;
-int (*SSL_CTX_set_ecdh_auto)(SSL_CTX *, int) = NULL;
+long (*SSL_CTX_set_ecdh_auto)(SSL_CTX *, int) = NULL;
 #else
 static const long Cryptography_HAS_SET_ECDH_AUTO = 1;
 #endif
